refactor: move WidgetItem out of message
