+++ src/main/java/org/runejs/client/message/inbound/widget/container/UpdateAllWidgetItemsInboundMessage.java
@@ -3,19 +3,6 @@ package org.runejs.client.message.inbound.widget.container;
 import org.runejs.client.message.InboundMessage;
 
 public class UpdateAllWidgetItemsInboundMessage implements InboundMessage {
-
-    public static class WidgetItem {
-        public final int index;
-        public final int id;
-        public final int amount;
-
-        public WidgetItem(int index, int id, int amount) {
-            this.index = index;
-            this.id = id;
-            this.amount = amount;
-        }
-    }
-
     public final int widgetId;
 
     public final int childId;
